refactor: move grid connector types to component subfolder
